template expansion is now functional
Also add additionalParameters to the template of the commit status
resource definition

Missing: good error reporting for the commit status resource user.
Right now, they have to inspect the logs. Is that good enough? Do we
want to report the malfunctioning commit status POST on the commit
status resource itself?